fix: ensure Action Menu Item with [href] close the menu
